+++ lib/material_palette.dart
@@ -28,8 +28,8 @@ export 'package:material_palette/src/shaders/radial_voronoi_gradient_shader_fill
 export 'package:material_palette/src/shaders/radial_voronoise_gradient_shader_fill.dart';
 // Special fills
 export 'package:material_palette/src/shaders/marble_smear_shader_fill.dart';
-export 'package:material_palette/src/shaders/liquid_metal_shader_fill.dart';
-export 'package:material_palette/src/shaders/layered_metal_shader_fill.dart';
+export 'package:material_palette/src/shaders/liquid_patina_shader_fill.dart';
+export 'package:material_palette/src/shaders/metal_smoke_shader_fill.dart';
 // Fur fills
 export 'package:material_palette/src/shaders/fur_planar_shader_fill.dart';
 
@@ -49,7 +49,7 @@ export 'package:material_palette/src/shaders/tappable_slurp_shader_wrap.dart';
 // Fur wraps
 export 'package:material_palette/src/shaders/fur_planar_mask_shader_wrap.dart';
 // Turbulence wraps
-export 'package:material_palette/src/shaders/turbulence_mask_shader_wrap.dart';
+export 'package:material_palette/src/shaders/turbulence_shader_wrap.dart';
 // Dither wraps
 export 'package:material_palette/src/shaders/dither_shader_wrap.dart';
 // Peel wraps